Migrate haiku_image_classif.py to notebook format: resnet_haiku.ipynb
Follows a similar pattern to the resnet_flax.ipynb notebook, but using
haiku instead of flax. Also gets SOTA results on CIFAR-10 and takes
around 13 minutes to run on a single GPU (same as the flax version).

Interestingly, this notebook is shorter than the flax one since we
don't need to define Resnets, we can just import them from Haiku.